♻️ refactor(server): generalize SeekableAudioSource → io/SeekableSource
