+++ server/src/jvmMain/kotlin/com/calypsan/listenup/server/io/SeekableSource.jvm.kt
@@ -1,14 +1,14 @@
-package com.calypsan.listenup.server.embeddedmeta
+package com.calypsan.listenup.server.io
 
 import java.io.RandomAccessFile
 import kotlinx.io.files.Path
 
-internal actual fun defaultSeekableSource(path: Path): SeekableAudioSource =
+internal actual fun openSeekableSource(path: Path): SeekableSource =
     RandomAccessFileSource(RandomAccessFile(path.toString(), "r"))
 
 private class RandomAccessFileSource(
     private val raf: RandomAccessFile,
-) : SeekableAudioSource {
+) : SeekableSource {
     override val length: Long get() = raf.length()
 
     override fun position(): Long = raf.filePointer
